langchain-api/requirements.txt
Flask
python-dotenv
langchain
openai
langchain-openai
